Type casting bug fixed

+++ app/src/main/java/com/axiom/atom/engine/graphics/gles2d/Texture.java
@@ -35,7 +35,7 @@ public class Texture implements GLESObject {
         //----------------------------------------------------------------
         // Загружаем текстуру если она еще не была загружна
         //---------------------------------------------------------------
-        Texture texture = loadedTextures.get(resource);
+        Texture texture = loadedTextures.get((long)resource);
         if (texture==null) {
             // Текстуру загружаем только один раз
             texture = new Texture(resources, resource);
